pyproject.toml
[tool.poetry]
name = "video-to-gif"
version = "0.1.0"
description = "동영상을 GIF로 변환하는 웹 애플리케이션"
authors = ["koa <your_email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.110.0"
uvicorn = "^0.27.1"
python-multipart = "^0.0.9"
ffmpeg-python = "^0.2.0"
pydantic-settings = "^2.2.1"
python-dotenv = "^1.0.1"
aiofiles = "^23.2.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.1.0"
flake8 = "^7.0.0"
mypy = "^1.5.1"
isort = "^5.12.0"
pre-commit = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py312"]  # 가장 최신 지원 버전인 3.12로 변경

[tool.isort]
profile = "black"
line_length = 88
